Move HnswGraph to hnsw package
