+++ lucene/core/src/java/org/apache/lucene/util/hnsw/OnHeapHnswGraph.java
@@ -22,7 +22,6 @@ import static org.apache.lucene.search.DocIdSetIterator.NO_MORE_DOCS;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import org.apache.lucene.index.HnswGraph;
 import org.apache.lucene.util.ArrayUtil;
 
 /**
